+++ src/index.ts
@@ -29,7 +29,7 @@ export function createApiV1(options: IOptions = createDefaultOptions()): RestApi
 export function createApiV2(options: IOptions = createDefaultOptions(ApiVersions.V2)): RestApiV2 {
 	// V1 used because organization is not supported in V2 yet
 	const { accountOptions, organizationOptions } = createApiOrgAndAccountOptions(options, ApiVersions.V1);
-	return new RestApiV2(accountOptions, organizationOptions);
+	return new RestApiV2({ ...accountOptions, version: ApiVersions.V2 }, { ...organizationOptions, version: ApiVersions.V2 });
 }
 
 /** @deprecated, use createApiV1, or createApiV2 functions instead */